requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-gis==1.0
drf-nested-routers==0.93.4
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
